+++ NEWS.md
@@ -1,6 +1,8 @@
 # v0.1.2.001
 Major changes:
 - New functions `dodgr_fundamental_cycles` and `dodgr_full_cycles`
+- New function `dodgr_sflines_to_poly` to convert `sf` collections of
+  `LINESTRING` object into corresponding enclosed `POLYGON` objects.
 
 
 # v0.1.2